pyproject.toml
[project]
name = "oi-bot"
version = "3.1.0"
description = "This chatbot provides match information for OIer and ACMer, based on the Nonebot2 framework."
readme = "README.md"
requires-python = ">=3.12, <4.0"
authors = [{ name = "talentestors", email = "talentestors@outlook.com" }]

license = "AGPL-3.0-or-later"

dependencies = [
    "nonebot2[httpx,fastapi]>=2.4.2",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-adapter-console>=0.6.0",
    "nonechat==0.3.0",
    "nonebot-plugin-oi-helper>=1.1.0",
    "nonebot-plugin-alconna>=0.57.6",
    "nonebot-plugin-whateat-pic>=1.5.7",
    "nonebot-plugin-localstore>=0.7.4",
]

[dependency-groups]
dev = [
    "nb-cli>=1.4.2",
    "pip>=25.1.0",
    "ruff>=0.14.11",
]
test = [
    "nonebug>=0.4.3",
    "pytest-asyncio>=0.26.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
